# backend/docs/swagger.yaml
basePath: /todos
host: localhost:8080
info:
  contact:
    email: frloit01@hs-esslingen.de
    name: Friedrich Lohrmann
  description: Backend endpoints created for distributed systems lab
  title: Todo API
  version: "1.0"
paths:
  /todos/:
    get:
      description: GET method for all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GET method for all tasks
      tags:
      - todos
  /todos/{task}:
    delete:
      description: DELETE method for one task, fetched by title
      parameters:
      - description: task title
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: DELETE method for one task
      tags:
      - todos
    get:
      description: GET method for one task, fetched by title
      parameters:
      - description: task title
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GET method for one task
      tags:
      - todos
    post:
      description: POST method for a new task
      parameters:
      - description: task title
        in: path
        name: task
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: POST method for a new task
      tags:
      - todos
swagger: "2.0"